requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
dj-database-url==2.1.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
gunicorn==21.2.0
